refactor: 클래스 및 메서드명 변경
