+++ src/main/java/ajaajaja/debugging_rounge/feature/auth/domain/repository/BlacklistedRefreshTokenRepository.java
@@ -1,8 +1,8 @@
 package ajaajaja.debugging_rounge.feature.auth.domain.repository;
 
-import ajaajaja.debugging_rounge.feature.auth.domain.BlackRefreshToken;
+import ajaajaja.debugging_rounge.feature.auth.domain.BlacklistedRefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlackRefreshTokenRepository extends JpaRepository<BlackRefreshToken, Long> {
+public interface BlacklistedRefreshTokenRepository extends JpaRepository<BlacklistedRefreshToken, Long> {
     Boolean existsByRefreshToken(String refreshToken);
 }
